Refactor src folder structure

+++ src/components/SceneTree.cpp
@@ -1,11 +1,9 @@
-#include "moc_scene-tree.cpp"
+#include "SceneTree.hpp"
 
-#include <QSizePolicy>
 #include <QScrollBar>
-#include <QDropEvent>
-#include <QPushButton>
 #include <QTimer>
-#include <cmath>
+
+#include "moc_SceneTree.cpp"
 
 SceneTree::SceneTree(QWidget *parent_) : QListWidget(parent_)
 {
@@ -79,7 +77,7 @@ void SceneTree::resizeEvent(QResizeEvent *event)
 		}
 
 		int wid = contentsRect().width() - scrollWid - 1;
-		int items = (int)std::ceil((float)wid / (float)maxWidth);
+		int items = (int)std::ceil((float)wid / maxWidth);
 		int itemWidth = wid / items;
 
 		setGridSize(QSize(itemWidth, itemHeight));
@@ -123,14 +121,14 @@ void SceneTree::dropEvent(QDropEvent *event)
 			setVerticalScrollBarPolicy(Qt::ScrollBarAlwaysOn);
 		}
 
-		float wid = float(contentsRect().width() - scrollWid - 1);
+		float wid = contentsRect().width() - scrollWid - 1;
 
 		QPoint point = event->position().toPoint();
 
-		int x = int((float)point.x() / wid * std::ceil(wid / (float)maxWidth));
+		int x = (float)point.x() / wid * std::ceil(wid / maxWidth);
 		int y = (point.y() + firstItemY) / itemHeight;
 
-		int r = x + y * (int)std::ceil(wid / (float)maxWidth);
+		int r = x + y * std::ceil(wid / maxWidth);
 
 		QListWidgetItem *item = takeItem(selectedIndexes()[0].row());
 		insertItem(r, item);
@@ -163,15 +161,15 @@ void SceneTree::RepositionGrid(QDragMoveEvent *event)
 		setVerticalScrollBarPolicy(Qt::ScrollBarAlwaysOn);
 	}
 
-	float wid = float(contentsRect().width() - scrollWid - 1);
+	float wid = contentsRect().width() - scrollWid - 1;
 
 	if (event) {
 		QPoint point = event->position().toPoint();
 
-		int x = int((float)point.x() / wid * std::ceil(wid / (float)maxWidth));
+		int x = (float)point.x() / wid * std::ceil(wid / maxWidth);
 		int y = (point.y() + firstItemY) / itemHeight;
 
-		int r = x + y * (int)std::ceil(wid / (float)maxWidth);
+		int r = x + y * std::ceil(wid / maxWidth);
 		int orig = selectedIndexes()[0].row();
 
 		for (int i = 0; i < count(); i++) {
@@ -184,8 +182,8 @@ void SceneTree::RepositionGrid(QDragMoveEvent *event)
 
 			int off = (i >= r ? 1 : 0) - (i > orig && i > r ? 1 : 0) - (i > orig && i == r ? 2 : 0);
 
-			int xPos = (i + off) % (int)std::ceil(wid / (float)maxWidth);
-			int yPos = (i + off) / (int)std::ceil(wid / (float)maxWidth);
+			int xPos = (i + off) % (int)std::ceil(wid / maxWidth);
+			int yPos = (i + off) / (int)std::ceil(wid / maxWidth);
 			QSize g = gridSize();
 
 			QPoint position(xPos * g.width(), yPos * g.height());
@@ -200,8 +198,8 @@ void SceneTree::RepositionGrid(QDragMoveEvent *event)
 
 			QModelIndex index = indexFromItem(wItem);
 
-			int xPos = i % (int)std::ceil(wid / (float)maxWidth);
-			int yPos = i / (int)std::ceil(wid / (float)maxWidth);
+			int xPos = i % (int)std::ceil(wid / maxWidth);
+			int yPos = i / (int)std::ceil(wid / maxWidth);
 			QSize g = gridSize();
 
 			QPoint position(xPos * g.width(), yPos * g.height());
